Migrated camera engine methods to security_engine
