+++ MEP_Engine/MEP_Engine.csproj
@@ -77,7 +77,6 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Compute\Error.cs" />
     <Compile Include="Compute\PlumbingFixtureWaterDemandPerDay.cs" />
     <Compile Include="Create\CoolingCoil.cs" />
     <Compile Include="Create\ElectricalConnector.cs" />
