+++ MEP_Engine/Query/Geometry.cs
@@ -43,32 +43,5 @@ namespace BH.Engine.MEP
         }
 
         /***************************************************/
-
-        [Description("Gets the camera's geometry as a closed ICurve cone-shape. Method required for automatic display in UI packages.")]
-        [Input("cameraDevice", "Camera to get the ICurve from.")]
-        [Output("icurve", "The geometry of the Camera.")]
-        public static ICurve Geometry(this CameraDevice cameraDevice)
-        {
-            Vector direction = BH.Engine.Geometry.Create.Vector(
-                BH.Engine.Geometry.Create.Point(cameraDevice.EyePosition.X, cameraDevice.EyePosition.Y, 0),
-                BH.Engine.Geometry.Create.Point(cameraDevice.TargetPosition.X, cameraDevice.TargetPosition.Y, 0)).Normalise();
-
-            Vector perpendicular = direction.Rotate((Math.PI / 180) * 90, Vector.ZAxis);
-
-            List<Point> vertices = new List<Point>();
-            Point point1 = cameraDevice.EyePosition.Clone();
-            Point point2 = cameraDevice.TargetPosition.Clone().Translate(perpendicular * (cameraDevice.HorizontalFieldOfView / 2));
-            Point point3 = cameraDevice.TargetPosition.Clone().Translate(perpendicular * ((cameraDevice.HorizontalFieldOfView / 2)) * -1);
-            vertices.Add(point1);
-            vertices.Add(point2);
-            vertices.Add(point3);
-            vertices.Add(point1);
-
-            Polyline polyline = BH.Engine.Geometry.Create.Polyline(vertices);
-
-            return polyline;
-        }
-
-        /***************************************************/
     }
 }
