+++ Security_Engine/Query/Centreline.cs
@@ -23,23 +23,23 @@
 using System.ComponentModel;
 
 using BH.oM.Geometry;
-using BH.oM.MEP.System;
-using BH.oM.MEP.Fixtures;
+using BH.oM.Security.Elements;
 using BH.oM.Reflection.Attributes;
 
-namespace BH.Engine.MEP
+namespace BH.Engine.Security
 {
     public static partial class Query
     {
         /***************************************************/
         /****              Public Methods               ****/
         /***************************************************/
-        [Description("Returns the centreline of any IFlow object as the line between the StartPoint and EndPoint. No offsets or similar is accounted for.")]
-        [Input("flowObj", "The IFlow object to get the centreline from.")]
-        [Output("centreline", "The centreline of the IFlow object.")]
-        public static Line Centreline(this IFlow flowObj)
+
+        [Description("Returns the centreline of a CameraDevice object.")]
+        [Input("cameraDevice", "The CameraDevice object to get the centreline from.")]
+        [Output("centreline", "The centreline of the CameraDevice object.")]
+        public static Line Centreline(this CameraDevice cameraDevice)
         {
-            return new Line { Start = flowObj.StartPoint, End = flowObj.EndPoint };
+            return new Line { Start = cameraDevice.EyePosition, End = cameraDevice.TargetPosition };
         }
 
         /***************************************************/
